requirements.txt
pip>8
scrapy==1.2
html2text
pydub
transliterate
pymongo
PyYAML
bs4
slugify
requests
mongokit